Move config functions to config module and reduce codecov target
- Move validate_yaml_syntax and get_meshtastic_config_value functions from cli.py to config.py
- Remove duplicate get_meshtastic_config_value functions from matrix_utils.py and meshtastic_utils.py
- Update all imports and function calls to use centralized config functions
- Reduce configuration component codecov target from 70% to 50%
- Fix runtime test to use new function signature with config parameter
