+++ src/mmrelay/cli.py
@@ -12,7 +12,7 @@ from yaml.loader import SafeLoader
 
 # Import version from package
 from mmrelay import __version__
-from mmrelay.config import get_config_paths
+from mmrelay.config import get_config_paths, validate_yaml_syntax
 from mmrelay.constants.app import WINDOWS_PLATFORM
 from mmrelay.constants.config import (
     CONFIG_KEY_ACCESS_TOKEN,
@@ -127,110 +127,6 @@ def print_version():
     """
     print(f"MMRelay v{__version__}")
 
-
-def validate_yaml_syntax(config_content, config_path):
-    """
-    Validate YAML syntax and provide detailed error messages for common issues.
-
-    Args:
-        config_content (str): The raw YAML content
-        config_path (str): Path to the config file for error reporting
-
-    Returns:
-        tuple: (is_valid, error_message, parsed_config)
-    """
-    import re
-
-    lines = config_content.split("\n")
-
-    # Check for common YAML syntax issues
-    syntax_issues = []
-
-    for line_num, line in enumerate(lines, 1):
-        # Skip empty lines and comments
-        if not line.strip() or line.strip().startswith("#"):
-            continue
-
-        # Check for unclosed quotes
-        if line.count('"') % 2 != 0:
-            syntax_issues.append(
-                f"Line {line_num}: Unclosed double quote - {line.strip()}"
-            )
-        if line.count("'") % 2 != 0:
-            syntax_issues.append(
-                f"Line {line_num}: Unclosed single quote - {line.strip()}"
-            )
-
-        # Check for missing colons in key-value pairs
-        if ":" not in line and "=" in line:
-            syntax_issues.append(
-                f"Line {line_num}: Use ':' instead of '=' for YAML - {line.strip()}"
-            )
-
-        # Check for non-standard boolean values (style warning)
-        bool_pattern = r":\s*(yes|no|on|off|Yes|No|YES|NO)\s*$"
-        if re.search(bool_pattern, line):
-            match = re.search(bool_pattern, line)
-            non_standard_bool = match.group(1)
-            syntax_issues.append(
-                f"Line {line_num}: Style warning - Consider using 'true' or 'false' instead of '{non_standard_bool}' for clarity - {line.strip()}"
-            )
-
-    # Try to parse YAML and catch specific errors
-    try:
-        parsed_config = yaml.load(config_content, Loader=SafeLoader)
-        if syntax_issues:
-            # Separate warnings from errors
-            warnings = [issue for issue in syntax_issues if "Style warning" in issue]
-            errors = [issue for issue in syntax_issues if "Style warning" not in issue]
-
-            if errors:
-                return False, "\n".join(errors), None
-            elif warnings:
-                # Return success but with warnings
-                return True, "\n".join(warnings), parsed_config
-        return True, None, parsed_config
-    except yaml.YAMLError as e:
-        error_msg = f"YAML parsing error in {config_path}:\n"
-
-        # Extract line and column information if available
-        if hasattr(e, "problem_mark"):
-            mark = e.problem_mark
-            error_line = mark.line + 1
-            error_column = mark.column + 1
-            error_msg += f"  Line {error_line}, Column {error_column}: "
-
-            # Show the problematic line
-            if error_line <= len(lines):
-                problematic_line = lines[error_line - 1]
-                error_msg += f"\n  Problematic line: {problematic_line}\n"
-                error_msg += f"  Error position: {' ' * (error_column - 1)}^\n"
-
-        # Add the original error message
-        error_msg += f"  {str(e)}\n"
-
-        # Provide helpful suggestions based on error type
-        error_str = str(e).lower()
-        if "mapping values are not allowed" in error_str:
-            error_msg += "\n  Suggestion: Check for missing quotes around values containing special characters"
-        elif "could not find expected" in error_str:
-            error_msg += "\n  Suggestion: Check for unclosed quotes or brackets"
-        elif "found character that cannot start any token" in error_str:
-            error_msg += (
-                "\n  Suggestion: Check for invalid characters or incorrect indentation"
-            )
-        elif "expected <block end>" in error_str:
-            error_msg += (
-                "\n  Suggestion: Check indentation - YAML uses spaces, not tabs"
-            )
-
-        # Add syntax issues if found
-        if syntax_issues:
-            error_msg += "\n\nAdditional syntax issues found:\n" + "\n".join(
-                syntax_issues
-            )
-
-        return False, error_msg, None
 
 
 def check_config(args=None):
