+++ src/mmrelay/matrix_utils.py
@@ -21,6 +21,7 @@ from nio import (
 from nio.events.room_events import RoomMemberEvent
 from PIL import Image
 
+from mmrelay.config import get_meshtastic_config_value
 from mmrelay.constants.config import (
     CONFIG_KEY_ACCESS_TOKEN,
     CONFIG_KEY_HOMESERVER,
@@ -49,35 +50,6 @@ from mmrelay.message_queue import get_message_queue, queue_message
 logger = get_logger(name="matrix_utils")
 
 
-def get_meshtastic_config_value(key, default=None, required=False):
-    """
-    Safely get a meshtastic configuration value with proper error handling.
-
-    Args:
-        key (str): Configuration key to retrieve
-        default: Default value if key is missing
-        required (bool): Whether this configuration is required
-
-    Returns:
-        The configuration value or default
-
-    Raises:
-        KeyError: If required=True and key is missing
-    """
-    try:
-        return config["meshtastic"][key]
-    except KeyError:
-        if required:
-            logger.error(
-                f"Missing required configuration: meshtastic.{key}\n"
-                f"Please add '{key}: {default if default is not None else 'VALUE'}' to your meshtastic section in config.yaml\n"
-                f"Run 'mmrelay --check-config' to validate your configuration."
-            )
-            raise KeyError(
-                f"Required configuration 'meshtastic.{key}' is missing. "
-                f"Add '{key}: {default if default is not None else 'VALUE'}' to your meshtastic section."
-            ) from None
-        return default
 
 
 def _get_msgs_to_keep_config():
@@ -790,7 +762,7 @@ async def send_reply_to_meshtastic(
     meshtastic_channel = room_config["meshtastic_channel"]
 
     if get_meshtastic_config_value(
-        "broadcast_enabled", DEFAULT_BROADCAST_ENABLED, required=False
+        config, "broadcast_enabled", DEFAULT_BROADCAST_ENABLED, required=False
     ):
         try:
             # Create mapping info once if storage is enabled
@@ -1075,7 +1047,7 @@ async def on_room_message(
             meshtastic_channel = room_config["meshtastic_channel"]
 
             if get_meshtastic_config_value(
-                "broadcast_enabled", DEFAULT_BROADCAST_ENABLED, required=False
+                config, "broadcast_enabled", DEFAULT_BROADCAST_ENABLED, required=False
             ):
                 meshtastic_logger.info(
                     f"Relaying reaction from remote meshnet {meshnet_name} to radio broadcast"
@@ -1150,7 +1122,7 @@ async def on_room_message(
             meshtastic_channel = room_config["meshtastic_channel"]
 
             if get_meshtastic_config_value(
-                "broadcast_enabled", DEFAULT_BROADCAST_ENABLED, required=False
+                config, "broadcast_enabled", DEFAULT_BROADCAST_ENABLED, required=False
             ):
                 meshtastic_logger.info(
                     f"Relaying reaction from {full_display_name} to radio broadcast"
@@ -1284,13 +1256,13 @@ async def on_room_message(
     # The lack of message_map storage just means no reaction bridging will occur.
     if not found_matching_plugin:
         if get_meshtastic_config_value(
-            "broadcast_enabled", DEFAULT_BROADCAST_ENABLED, required=False
+            config, "broadcast_enabled", DEFAULT_BROADCAST_ENABLED, required=False
         ):
             portnum = event.source["content"].get("meshtastic_portnum")
             if portnum == DETECTION_SENSOR_APP:
                 # If detection_sensor is enabled, forward this data as detection sensor data
                 if get_meshtastic_config_value(
-                    "detection_sensor", DEFAULT_DETECTION_SENSOR
+                    config, "detection_sensor", DEFAULT_DETECTION_SENSOR
                 ):
                     success = queue_message(
                         meshtastic_interface.sendData,
